Change ems() 'random' parameter default argument from NULL to ""
The ems() function does not currently behave correctly when the user does
not supply any arguments to the 'random' parameter. The default argument is
NULL. The intention is that this will be interpreted as all factors being
fixed. Instead, it is interpreted as all factors being random. I had not
noticed this before because I pretty much only use ems() as part of the
underlying machinery to my PANGEA shiny app, and the way the function is
called there, an argument is always explicitly passed to 'random' (i.e., if
the user selects all factors fixed, ems is called with random="", rather than
omitting the random parameter). Anyway, the solution appears to be simple:
just change the default argument from NULL to "". That is what this commit
does. It appears to work fine now with this change.